# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"The grand-game API","description":"This API serves the game","version":"0.0.1"},"host":"localhost:8080","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/advantages":{"get":{"tags":["game"],"summary":"list_advantages game","description":"Returns all the advantages and their information","operationId":"game#list_advantages","responses":{"200":{"description":"OK response.","schema":{"type":"array","items":{"$ref":"#/definitions/AdvantageResponse"}}},"404":{"description":"Not Found response.","schema":{"$ref":"#/definitions/GameListAdvantagesNotFoundResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/GameListAdvantagesInternalErrorResponseBody"}}},"schemes":["http"]}},"/competition_events/kinds":{"get":{"tags":["game"],"summary":"list_competition_event_kinds game","description":"Returns all the competition event kinds","operationId":"game#list_competition_event_kinds","responses":{"200":{"description":"OK response.","schema":{"type":"array","items":{"$ref":"#/definitions/CompetitionEventDescriptionResponse"}}},"404":{"description":"Not Found response.","schema":{"$ref":"#/definitions/GameListCompetitionEventKindsNotFoundResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/GameListCompetitionEventKindsInternalErrorResponseBody"}}},"schemes":["http"]}},"/participants":{"get":{"tags":["game"],"summary":"list_participants game","description":"Returns all the participants and their information","operationId":"game#list_participants","responses":{"200":{"description":"OK response.","schema":{"type":"array","items":{"$ref":"#/definitions/ParticipantResponse"}}},"404":{"description":"Not Found response.","schema":{"$ref":"#/definitions/GameListParticipantsNotFoundResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/GameListParticipantsInternalErrorResponseBody"}}},"schemes":["http"]}}},"definitions":{"AdvantageResponse":{"title":"AdvantageResponse","type":"object","properties":{"Description":{"type":"string","description":"Description of advantage","example":"Id magnam nam consequatur ea totam."},"Events":{"type":"array","items":{"type":"string","example":"Consectetur aut sunt ratione illo voluptates natus."},"description":"Events that allow this advantage","example":["Omnis est cum dolore ut.","In omnis non.","Consequatur voluptatum.","Iusto aperiam nihil animi."]},"Id":{"type":"integer","description":"Unique advantage ID","example":8480116710669633731,"format":"int64"},"Name":{"type":"string","description":"Name of advantage","example":"Voluptatem ipsam est fugiat."},"Tier":{"type":"string","description":"Tier of advantage","example":"Busted","enum":["Slight","Major","Busted"]}},"example":{"Description":"Maiores vitae odit rerum.","Events":["Perspiciatis aut odio dolor officia est.","Sequi qui."],"Id":8956198482236671647,"Name":"Quam facere.","Tier":"Busted"},"required":["Id","Name","Description","Events","Tier"]},"CompetitionEventDescriptionResponse":{"title":"CompetitionEventDescriptionResponse","type":"object","properties":{"Description":{"type":"string","description":"Description of competition event","example":"Et aut odit."},"Id":{"type":"integer","description":"Unique competition event description ID","example":454275695579401491,"format":"int64"},"Name":{"type":"string","description":"Name of competition event","example":"Atque voluptatibus."}},"example":{"Description":"Et quis optio sequi voluptatem consequuntur ut.","Id":1332868130237203961,"Name":"Sunt doloremque."},"required":["Id","Name","Description"]},"GameListAdvantagesInternalErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"list_advantages_internal_error_response_body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"GameListAdvantagesNotFoundResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"list_advantages_not_found_response_body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"GameListCompetitionEventKindsInternalErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"list_competition_event_kinds_internal_error_response_body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"GameListCompetitionEventKindsNotFoundResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"list_competition_event_kinds_not_found_response_body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"GameListParticipantsInternalErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"list_participants_internal_error_response_body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"GameListParticipantsNotFoundResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"list_participants_not_found_response_body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"ParticipantResponse":{"title":"ParticipantResponse","type":"object","properties":{"Backstory":{"type":"string","description":"Backstory of participant","example":"Qui dignissimos voluptatem ipsam."},"Email":{"type":"string","description":"Email of participant","example":"Quo iure doloribus architecto cumque velit."},"Favor":{"type":"integer","description":"Influence of participant","example":2714003309803552099,"format":"int64"},"Gold":{"type":"integer","description":"Gold of participant","example":4731892199166666403,"format":"int64"},"Id":{"type":"integer","description":"Unique participant ID","example":8746297636993790074,"format":"int64"},"Kind":{"type":"string","description":"Kind of participant","example":"superadmin","enum":["player","admin","superadmin"]},"Name":{"type":"string","description":"Name of participant","example":"Non quia quia qui deleniti."},"Phone":{"type":"string","description":"Phone number of participant","example":"Voluptas officia nisi."},"Role":{"type":"string","description":"Role of participant","example":"Noble","enum":["King","King's Hand","Archvizier","Noble","Mercenary"]}},"example":{"Backstory":"Velit eos.","Email":"Itaque expedita ex.","Favor":8543632858255068945,"Gold":5419108875568623391,"Id":1466269278829255293,"Kind":"admin","Name":"Nulla harum qui nulla dicta.","Phone":"Consequatur quia harum.","Role":"King's Hand"},"required":["Id","Name","Email","Phone","Kind","Role","Backstory","Gold","Favor"]}}}